backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
sentence-transformers==3.0.0
faiss-cpu==1.8.0
numpy==1.26.4
scikit-learn==1.4.2
python-multipart==0.0.9
httpx==0.27.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
